Extract types from bundle client.
