+++ enterprise/internal/codeintel/bundles/database/mock_database_test.go
@@ -4,7 +4,7 @@ package database
 
 import (
 	"context"
-	client "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/bundles/client"
+	client "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/bundles/client_types"
 	"sync"
 )
